[refactor] 파일명 수정
- common 폴더 내부에 들어있어서 중복적인 의미 나타나 수정하였음
