+++ src/store/common/properties.ts
